Rename LocalAuthenticator to EmailAuthenticator.
