+++ packages/authentication/src/pre-hooks/index.ts
@@ -1,1 +1,1 @@
-export { validateLocalCredentialsFormat } from './validate-local-credentials-format.pre-hook';
+export { validateEmailCredentialsFormat } from './validate-email-credentials-format.pre-hook';
