+++ packages/examples/src/app/authentication/authenticator.service.ts
@@ -1,10 +1,10 @@
-import { LocalAuthenticatorService } from '@foal/authentication';
+import { EmailAuthenticatorService } from '@foal/authentication';
 import { Service } from '@foal/core';
 
 import { User, UserService } from '../shared';
 
 @Service()
-export class AuthenticatorService extends LocalAuthenticatorService<User> {
+export class AuthenticatorService extends EmailAuthenticatorService<User> {
   constructor(userService: UserService) {
     super(userService);
   }
